+++ src/terminal/Cell.h
@@ -134,6 +134,22 @@ class CONTOUR_PACKED Cell
 
     CellFlags styles() const noexcept;
 
+    void resetFlags() noexcept
+    {
+        if (extra_)
+            extra_->flags = CellFlags::None;
+    }
+
+    void setFlags(CellFlags flags, bool enable = true)
+    {
+        if (enable)
+            extra().flags = extra().flags | flags;
+        else
+            extra().flags = CellFlags(int(extra().flags) & ~int(flags));
+    }
+
+    void setGraphicsRendition(GraphicsRendition sgr) noexcept;
+
     Color underlineColor() const noexcept;
     void setUnderlineColor(Color color) noexcept;
     Color foregroundColor() const noexcept;
@@ -525,6 +541,45 @@ inline void Cell::setHyperlink(HyperlinkId _hyperlink)
     else if (extra_)
         extra_->hyperlink = {};
 }
+
+inline void Cell::setGraphicsRendition(GraphicsRendition _rendition) noexcept
+{
+    // TODO: optimize this as there are only 3 cases
+    // 1.) reset
+    // 2.) set some bits |=
+    // 3.) clear some bits &= ~
+    switch (_rendition)
+    {
+    case GraphicsRendition::Reset: extra().flags = {}; break;
+    case GraphicsRendition::Bold: extra().flags |= CellFlags::Bold; break;
+    case GraphicsRendition::Faint: extra().flags |= CellFlags::Faint; break;
+    case GraphicsRendition::Italic: extra().flags |= CellFlags::Italic; break;
+    case GraphicsRendition::Underline: extra().flags |= CellFlags::Underline; break;
+    case GraphicsRendition::Blinking: extra().flags |= CellFlags::Blinking; break;
+    case GraphicsRendition::Inverse: extra().flags |= CellFlags::Inverse; break;
+    case GraphicsRendition::Hidden: extra().flags |= CellFlags::Hidden; break;
+    case GraphicsRendition::CrossedOut: extra().flags |= CellFlags::CrossedOut; break;
+    case GraphicsRendition::DoublyUnderlined: extra().flags |= CellFlags::DoublyUnderlined; break;
+    case GraphicsRendition::CurlyUnderlined: extra().flags |= CellFlags::CurlyUnderlined; break;
+    case GraphicsRendition::DottedUnderline: extra().flags |= CellFlags::DottedUnderline; break;
+    case GraphicsRendition::DashedUnderline: extra().flags |= CellFlags::DashedUnderline; break;
+    case GraphicsRendition::Framed: extra().flags |= CellFlags::Framed; break;
+    case GraphicsRendition::Overline: extra().flags |= CellFlags::Overline; break;
+    case GraphicsRendition::Normal: extra().flags &= ~(CellFlags::Bold | CellFlags::Faint); break;
+    case GraphicsRendition::NoItalic: extra().flags &= ~CellFlags::Italic; break;
+    case GraphicsRendition::NoUnderline:
+        extra().flags &= ~(CellFlags::Underline | CellFlags::DoublyUnderlined | CellFlags::CurlyUnderlined
+                           | CellFlags::DottedUnderline | CellFlags::DashedUnderline);
+        break;
+    case GraphicsRendition::NoBlinking: extra().flags &= ~CellFlags::Blinking; break;
+    case GraphicsRendition::NoInverse: extra().flags &= ~CellFlags::Inverse; break;
+    case GraphicsRendition::NoHidden: extra().flags &= ~CellFlags::Hidden; break;
+    case GraphicsRendition::NoCrossedOut: extra().flags &= ~CellFlags::CrossedOut; break;
+    case GraphicsRendition::NoFramed: extra().flags &= ~CellFlags::Framed; break;
+    case GraphicsRendition::NoOverline: extra().flags &= ~CellFlags::Overline; break;
+    }
+}
+
 // }}}
 
 } // namespace terminal